fix: crash when using column detection

+++ app/libs/image_processing/image_processing.cpp
@@ -196,7 +196,7 @@ extern "C" {
 
 
 
-    jint Java_com_googlecode_tesseract_android_NativeBinding_nativeAnalyseLayout(JNIEnv *env, jobject thiz, jint nativePix) {
+    jint Java_com_googlecode_tesseract_android_NativeBinding_nativeAnalyseLayout(JNIEnv *env, jobject thiz, jlong nativePix) {
         LOGV(__FUNCTION__);
         Pix *pixOrg = (PIX *) nativePix;
         Pix* pixTextlines = NULL;
